ConnectionPool: exchange connections between pools
Now ConnectionPool-s (who talk to the same Host from different
threads) can exchange connections. So if one pool establishes
a connection to a wrong shard, it can put this Connection into
a "marketplace" where it can be picked by other pool that happens
to need it.

The side effect is that excess of connections is sometimes opened
and it's not easy to determine the right moment to close them. So
excess connections are kept unused until ConnectionPool-s are
closed. The extra connections are visible in netstat (or
system.clients) but they are not used for queries.

Also, default reconnection policy is changed to "Constant" which
indicates that client software should consider configuring it this
way (it's overridden by cluster config though, so it's just a hint
with no effect). Choice of right RP significantly speeds up the
process of establishing a Session.