Fixed crash when using marketplace.
Non-advanced shard-awarness is using marketplace mechanism,
to let threads give out connections to other threads.
This was causing segmentation faults because using libuv's event loop
apis from a different thread than the loop is running on is
unsupported. Unfortunately libuv doesn't have any mechanism to move
handle to a different event loop, so I had to resort to a dirty hack:
extract fd from handle, dup() this fd, and then reconstruct object in
a different event loop. This seems to work, altough it is very fragile,
because changes to Connection/Socket classes are likely to break this fix.