Cluster: do not prolong the lifetime of Cluster for PrepareHostHandler
It was supposed to fix a crash when "host up" event interleaves with
the destruction of Cluster object, but it introduced another bug, which
can be observed e.g. in ControlConnectionTests.Integration_Cassandra_FullOutage
integration test. The bug is that `EventLoop` still has working "timer handle"
while being destroyed and, in turn, segfaults.