Rename static library target
Building the unit tests on Windows failed at the linking stage:

	LINK : fatal error LNK1104: cannot open file 'scylla-cpp-driver_static.lib'

This is caused by the tests being forced to use the static library on
Windows but link against `${PROJECT_LIB_NAME_TARGET}` which is set to
`scylla-cpp-driver_static` in the main `CMakeLists.txt` while the target
was actually still called `cassandra_static`.

Renamed the static library target to be consistent with the dynamic
library target and `${PROJECT_LIB_NAME_TARGET}` so the tests can be
successfully linked on Windows.