Use inet_pton instead of inet_addr
Compiling with MSVC produces the following warning:

	warning C4996: 'inet_addr': Use inet_pton() or InetPton() instead or define _WINSOCK_DEPRECATED_NO_WARNINGS to disable deprecated API warnings

Instead of just ignoring the deprecation warnings replacing `inet_addr`
is straight-forward here.

`inet_pton` works well as a cross-platform alternative. The difference
in parsing[^1] are acceptable since this is only used in test code where
it won't lead to a chance in behavior since the only used value is a
fixed string of "127.0.0.1" in `bridge.cpp`.

[^1]: https://man7.org/linux/man-pages/man3/inet_pton.3.html#NOTES